LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/unicode-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/unicode' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
Web Socket unicode message test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Send UNICODE_HELLO.
PASS data is UNICODE_GOODBYE
Closed.
PASS successfullyParsed is true

TEST COMPLETE
